+++ Dockerfile
@@ -3,5 +3,6 @@ FROM continuumio/miniconda3
 ADD ./conda.yaml .
 RUN conda env update -n base -f conda.yaml
 WORKDIR /app
+ENV PYTHONPATH "${PYTHONPATH}:/app"
 COPY . .
 CMD ["bash"]
