+++ api/jobs.py
@@ -237,110 +237,6 @@ class Dispatcher:
         return self.run()
 
 
-#################################
-# k8s
-
-
-def serialize_func(func):
-    code = base64.b64encode(pickle.dumps(func)).decode("utf-8")
-    return code
-
-
-def deserialize_func(code):
-    func = pickle.loads(base64.b64decode(code))
-    return func
-
-
-def random_string(length):
-    letters = string.ascii_lowercase
-    return ''.join(random.choice(letters) for i in range(length))
-
-
-def run_cmd(cmd):
-    result = subprocess.run(cmd, check=True, shell=True, stdout=subprocess.PIPE).stdout
-    return result
-
-
-def run_k8s(func, image="jobs"):
-    job_template = """apiVersion: batch/v1
-kind: Job
-metadata:
-  name: {{name}}
-spec:
-  template:
-    spec:
-      containers:
-      - name: {{name}}
-        image: {{image}}
-        imagePullPolicy: Never
-        command:
-        - python
-        - -c
-        - |
-          import dill as pickle
-          import base64
-          import json
-          func = pickle.loads(base64.b64decode("{{code}}"))
-          func()
-      restartPolicy: Never
-  backoffLimit: 1
-"""
-
-    code = serialize_func(func)
-    t = jinja2.Template(job_template)
-    s = random_string(5)
-    j = t.render(name=f"job-{s}", code=code, image=image)
-    subprocess.run("kubectl apply -f -",
-                   input=j.encode("utf-8"),
-                   check=True)
-    return f"job-{s}"
-
-
-def wait_k8s(job_name, timeout=None):
-    get_job = f"kubectl get job -o json {job_name}"
-
-    # Wait until the whole job is finished:
-    start = time.time()
-    while True:
-        current = time.time()
-        if (timeout is not None) and (current - start) > timeout:
-            return None
-        result = json.loads(run_cmd(get_job))
-        if "succeeded" not in result["status"]:
-            continue
-        if result["status"]["succeeded"] == 1:
-            break
-        time.sleep(1)
-
-    # Collect the names of the pods:
-    get_pods = f"kubectl get pods --selector=job-name={job_name} --output=json"
-    pods = json.loads(run_cmd(get_pods))
-    pod_names = [p["metadata"]["name"] for p in pods["items"]]
-
-    logs = {}
-    for pod_name in pod_names:
-        logs[pod_name] = json.loads(run_cmd(f"kubectl logs {pod_name}"))
-
-    return logs
-
-
-def map_k8s(func, iterable, image="jobs", timeout=None, wait=True, verbose=False):
-    thunks = [lambda arg=i: print(json.dumps(func(arg))) for i in iterable]
-    job_names = [run_k8s(thunk, image=image) for thunk in thunks]
-
-    if wait:
-        if verbose:
-            results = {j: wait_k8s(j, timeout=timeout) for j in job_names}
-        else:
-            results = [list(wait_k8s(j, timeout=timeout).values())[0] for j in job_names]
-        return results
-    else:
-        return job_names
-
-
-#################################
-
-
 if __name__=="__main__":
     #print("TEST", "test3", test3())
     print("TEST", "test4", test4())
